refactor: move SysCatalogReaderImpl::read_xxx to individual files for later refactoring
