+++ src/frontend/src/catalog/system_catalog/pg_catalog/pg_attribute.rs
@@ -12,9 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use risingwave_common::types::DataType;
+use itertools::Itertools;
+use risingwave_common::error::Result;
+use risingwave_common::row::OwnedRow;
+use risingwave_common::types::{DataType, ScalarImpl};
 
-use crate::catalog::system_catalog::SystemCatalogColumnsDef;
+use crate::catalog::system_catalog::{SysCatalogReaderImpl, SystemCatalogColumnsDef};
 
 /// The catalog `pg_attribute` stores information about table columns. There will be exactly one
 /// `pg_attribute` row for every column in every table in the database. (There will also be
@@ -34,3 +37,58 @@ pub const PG_ATTRIBUTE_COLUMNS: &[SystemCatalogColumnsDef<'_>] = &[
     (DataType::Boolean, "attisdropped"),
     (DataType::Int32, "atttypmod"),
 ];
+
+impl SysCatalogReaderImpl {
+    pub fn read_pg_attribute(&self) -> Result<Vec<OwnedRow>> {
+        let reader = self.catalog_reader.read_guard();
+        let schemas = reader.iter_schemas(&self.auth_context.database)?;
+
+        Ok(schemas
+            .flat_map(|schema| {
+                let view_rows = schema.iter_view().flat_map(|view| {
+                    view.columns.iter().enumerate().map(|(index, column)| {
+                        OwnedRow::new(vec![
+                            Some(ScalarImpl::Int32(view.id as i32)),
+                            Some(ScalarImpl::Utf8(column.name.clone().into())),
+                            Some(ScalarImpl::Int32(column.data_type().to_oid())),
+                            Some(ScalarImpl::Int16(column.data_type().type_len())),
+                            Some(ScalarImpl::Int16(index as i16 + 1)),
+                            Some(ScalarImpl::Bool(false)),
+                            Some(ScalarImpl::Bool(false)),
+                            // From https://www.postgresql.org/docs/current/catalog-pg-attribute.html
+                            // The value will generally be -1 for types that do not need
+                            // `atttypmod`.
+                            Some(ScalarImpl::Int32(-1)),
+                        ])
+                    })
+                });
+
+                schema
+                    .iter_valid_table()
+                    .flat_map(|table| {
+                        table
+                            .columns()
+                            .iter()
+                            .enumerate()
+                            .filter(|(_, column)| !column.is_hidden())
+                            .map(|(index, column)| {
+                                OwnedRow::new(vec![
+                                    Some(ScalarImpl::Int32(table.id.table_id() as i32)),
+                                    Some(ScalarImpl::Utf8(column.name().into())),
+                                    Some(ScalarImpl::Int32(column.data_type().to_oid())),
+                                    Some(ScalarImpl::Int16(column.data_type().type_len())),
+                                    Some(ScalarImpl::Int16(index as i16 + 1)),
+                                    Some(ScalarImpl::Bool(false)),
+                                    Some(ScalarImpl::Bool(false)),
+                                    // From https://www.postgresql.org/docs/current/catalog-pg-attribute.html
+                                    // The value will generally be -1 for types that do not need
+                                    // `atttypmod`.
+                                    Some(ScalarImpl::Int32(-1)),
+                                ])
+                            })
+                    })
+                    .chain(view_rows)
+            })
+            .collect_vec())
+    }
+}
