+++ src/frontend/src/catalog/system_catalog/pg_catalog/pg_constraint.rs
@@ -14,10 +14,11 @@
 
 use std::sync::LazyLock;
 
+use risingwave_common::error::Result;
 use risingwave_common::row::OwnedRow;
 use risingwave_common::types::DataType;
 
-use crate::catalog::system_catalog::SystemCatalogColumnsDef;
+use crate::catalog::system_catalog::{SysCatalogReaderImpl, SystemCatalogColumnsDef};
 
 /// The catalog `pg_constraint` records information about table and index inheritance hierarchies.
 /// Ref: [`https://www.postgresql.org/docs/current/catalog-pg-constraint.html`]
@@ -54,3 +55,9 @@ pub static PG_CONSTRAINT_COLUMNS: LazyLock<Vec<SystemCatalogColumnsDef<'_>>> =
     });
 
 pub static PG_CONSTRAINT_DATA_ROWS: LazyLock<Vec<OwnedRow>> = LazyLock::new(Vec::new);
+
+impl SysCatalogReaderImpl {
+    pub fn read_constraint_info(&self) -> Result<Vec<OwnedRow>> {
+        Ok(PG_CONSTRAINT_DATA_ROWS.clone())
+    }
+}
