+++ src/frontend/src/catalog/system_catalog/pg_catalog/pg_conversion.rs
@@ -12,9 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use risingwave_common::error::Result;
+use risingwave_common::row::OwnedRow;
 use risingwave_common::types::DataType;
 
-use crate::catalog::system_catalog::SystemCatalogColumnsDef;
+use crate::catalog::system_catalog::{SysCatalogReaderImpl, SystemCatalogColumnsDef};
 
 /// The catalog `pg_conversion` describes encoding conversion functions.
 /// Reference: [`https://www.postgresql.org/docs/current/catalog-pg-conversion.html`]
@@ -29,3 +31,9 @@ pub const PG_CONVERSION_COLUMNS: &[SystemCatalogColumnsDef<'_>] = &[
     (DataType::Int32, "conproc"),
     (DataType::Boolean, "condefault"),
 ];
+
+impl SysCatalogReaderImpl {
+    pub(crate) fn read_conversion_info(&self) -> Result<Vec<OwnedRow>> {
+        Ok(vec![])
+    }
+}
