+++ src/frontend/src/catalog/system_catalog/pg_catalog/pg_enum.rs
@@ -12,9 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use risingwave_common::error::Result;
+use risingwave_common::row::OwnedRow;
 use risingwave_common::types::DataType;
 
-use crate::catalog::system_catalog::SystemCatalogColumnsDef;
+use crate::catalog::system_catalog::{SysCatalogReaderImpl, SystemCatalogColumnsDef};
 
 /// The `pg_enum` catalog contains entries showing the values and labels for each enum type.
 /// The internal representation of a given enum value is actually the OID of its associated row in
@@ -26,3 +28,9 @@ pub const PG_ENUM_COLUMNS: &[SystemCatalogColumnsDef<'_>] = &[
     (DataType::Float32, "enumsortorder"),
     (DataType::Varchar, "enumlabel"),
 ];
+
+impl SysCatalogReaderImpl {
+    pub fn read_enum_info(&self) -> Result<Vec<OwnedRow>> {
+        Ok(vec![])
+    }
+}
