+++ src/frontend/src/catalog/system_catalog/pg_catalog/pg_index.rs
@@ -14,9 +14,13 @@
 
 use std::sync::LazyLock;
 
-use risingwave_common::types::DataType;
+use itertools::Itertools;
+use risingwave_common::array::ListValue;
+use risingwave_common::error::Result;
+use risingwave_common::row::OwnedRow;
+use risingwave_common::types::{DataType, ScalarImpl};
 
-use crate::catalog::system_catalog::SystemCatalogColumnsDef;
+use crate::catalog::system_catalog::{SysCatalogReaderImpl, SystemCatalogColumnsDef};
 
 /// The catalog `pg_index` contains part of the information about indexes.
 /// Ref: [`https://www.postgresql.org/docs/current/catalog-pg-index.html`]
@@ -33,3 +37,31 @@ pub static PG_INDEX_COLUMNS: LazyLock<Vec<SystemCatalogColumnsDef<'_>>> = LazyLo
         (DataType::Varchar, "indpred"),
     ]
 });
+
+impl SysCatalogReaderImpl {
+    pub fn read_index_info(&self) -> Result<Vec<OwnedRow>> {
+        let reader = self.catalog_reader.read_guard();
+        let schemas = reader.iter_schemas(&self.auth_context.database)?;
+
+        Ok(schemas
+            .flat_map(|schema| {
+                schema.iter_index().map(|index| {
+                    OwnedRow::new(vec![
+                        Some(ScalarImpl::Int32(index.id.index_id() as i32)),
+                        Some(ScalarImpl::Int32(index.primary_table.id.table_id() as i32)),
+                        Some(ScalarImpl::Int16(index.original_columns.len() as i16)),
+                        Some(ScalarImpl::List(ListValue::new(
+                            index
+                                .original_columns
+                                .iter()
+                                .map(|index| Some(ScalarImpl::Int16(index.get_id() as i16 + 1)))
+                                .collect_vec(),
+                        ))),
+                        None,
+                        None,
+                    ])
+                })
+            })
+            .collect_vec())
+    }
+}
