+++ src/frontend/src/catalog/system_catalog/pg_catalog/pg_inherits.rs
@@ -14,10 +14,11 @@
 
 use std::sync::LazyLock;
 
+use risingwave_common::error::Result;
 use risingwave_common::row::OwnedRow;
 use risingwave_common::types::DataType;
 
-use crate::catalog::system_catalog::SystemCatalogColumnsDef;
+use crate::catalog::system_catalog::{SysCatalogReaderImpl, SystemCatalogColumnsDef};
 
 /// The catalog `pg_inherits` records information about table and index inheritance hierarchies.
 /// Ref: [`https://www.postgresql.org/docs/current/catalog-pg-inherits.html`]
@@ -31,3 +32,9 @@ pub const PG_INHERITS_COLUMNS: &[SystemCatalogColumnsDef<'_>] = &[
 ];
 
 pub static PG_INHERITS_DATA_ROWS: LazyLock<Vec<OwnedRow>> = LazyLock::new(Vec::new);
+
+impl SysCatalogReaderImpl {
+    pub fn read_inherits_info(&self) -> Result<Vec<OwnedRow>> {
+        Ok(PG_INHERITS_DATA_ROWS.clone())
+    }
+}
