+++ src/frontend/src/catalog/system_catalog/pg_catalog/pg_namespace.rs
@@ -12,9 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use risingwave_common::types::DataType;
+use itertools::Itertools;
+use risingwave_common::error::Result;
+use risingwave_common::row::OwnedRow;
+use risingwave_common::types::{DataType, ScalarImpl};
+use risingwave_pb::user::grant_privilege::Object;
 
-use crate::catalog::system_catalog::SystemCatalogColumnsDef;
+use crate::catalog::system_catalog::{
+    get_acl_items, SysCatalogReaderImpl, SystemCatalogColumnsDef,
+};
 
 /// The catalog `pg_namespace` stores namespaces. A namespace is the structure underlying SQL
 /// schemas: each namespace can have a separate collection of relations, types, etc. without name
@@ -26,3 +32,28 @@ pub const PG_NAMESPACE_COLUMNS: &[SystemCatalogColumnsDef<'_>] = &[
     (DataType::Int32, "nspowner"),
     (DataType::Varchar, "nspacl"),
 ];
+
+impl SysCatalogReaderImpl {
+    pub fn read_namespace(&self) -> Result<Vec<OwnedRow>> {
+        let schemas = self
+            .catalog_reader
+            .read_guard()
+            .get_all_schema_info(&self.auth_context.database)?;
+        let user_reader = self.user_info_reader.read_guard();
+        let users = user_reader.get_all_users();
+        let username_map = user_reader.get_user_name_map();
+        Ok(schemas
+            .iter()
+            .map(|schema| {
+                OwnedRow::new(vec![
+                    Some(ScalarImpl::Int32(schema.id as i32)),
+                    Some(ScalarImpl::Utf8(schema.name.clone().into())),
+                    Some(ScalarImpl::Int32(schema.owner as i32)),
+                    Some(ScalarImpl::Utf8(
+                        get_acl_items(&Object::SchemaId(schema.id), &users, username_map).into(),
+                    )),
+                ])
+            })
+            .collect_vec())
+    }
+}
