+++ src/frontend/src/catalog/system_catalog/pg_catalog/pg_roles.rs
@@ -12,9 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use risingwave_common::types::DataType;
+use itertools::Itertools;
+use risingwave_common::error::Result;
+use risingwave_common::row::OwnedRow;
+use risingwave_common::types::{DataType, ScalarImpl};
 
-use crate::catalog::system_catalog::SystemCatalogColumnsDef;
+use crate::catalog::system_catalog::{SysCatalogReaderImpl, SystemCatalogColumnsDef};
 
 /// The catalog `pg_roles` provides access to information about database roles. This is simply a
 /// publicly readable view of `pg_authid` that blanks out the password field.
@@ -30,3 +33,25 @@ pub const PG_ROLES_COLUMNS: &[SystemCatalogColumnsDef<'_>] = &[
     (DataType::Boolean, "rolcanlogin"),
     (DataType::Varchar, "rolpassword"),
 ];
+
+impl SysCatalogReaderImpl {
+    pub fn read_roles_info(&self) -> Result<Vec<OwnedRow>> {
+        let reader = self.user_info_reader.read_guard();
+        let users = reader.get_all_users();
+        Ok(users
+            .iter()
+            .map(|user| {
+                OwnedRow::new(vec![
+                    Some(ScalarImpl::Int32(user.id as i32)),
+                    Some(ScalarImpl::Utf8(user.name.clone().into())),
+                    Some(ScalarImpl::Bool(user.is_super)),
+                    Some(ScalarImpl::Bool(true)),
+                    Some(ScalarImpl::Bool(user.can_create_user)),
+                    Some(ScalarImpl::Bool(user.can_create_db)),
+                    Some(ScalarImpl::Bool(user.can_login)),
+                    Some(ScalarImpl::Utf8("********".into())),
+                ])
+            })
+            .collect_vec())
+    }
+}
