+++ src/frontend/src/catalog/system_catalog/pg_catalog/pg_settings.rs
@@ -14,10 +14,11 @@
 
 use std::sync::LazyLock;
 
+use risingwave_common::error::Result;
 use risingwave_common::row::OwnedRow;
 use risingwave_common::types::DataType;
 
-use crate::catalog::system_catalog::SystemCatalogColumnsDef;
+use crate::catalog::system_catalog::{SysCatalogReaderImpl, SystemCatalogColumnsDef};
 
 /// The catalog `pg_settings` stores settings.
 /// Ref: [`https://www.postgresql.org/docs/current/view-pg-settings.html`]
@@ -26,3 +27,9 @@ pub const PG_SETTINGS_COLUMNS: &[SystemCatalogColumnsDef<'_>] =
     &[(DataType::Varchar, "name"), (DataType::Varchar, "setting")];
 
 pub static PG_SETTINGS_DATA_ROWS: LazyLock<Vec<OwnedRow>> = LazyLock::new(Vec::new);
+
+impl SysCatalogReaderImpl {
+    pub fn read_settings_info(&self) -> Result<Vec<OwnedRow>> {
+        Ok(PG_SETTINGS_DATA_ROWS.clone())
+    }
+}
