+++ src/frontend/src/catalog/system_catalog/pg_catalog/pg_shadow.rs
@@ -14,9 +14,13 @@
 
 use std::sync::LazyLock;
 
-use risingwave_common::types::DataType;
+use itertools::Itertools;
+use risingwave_common::error::{ErrorCode, Result};
+use risingwave_common::row::OwnedRow;
+use risingwave_common::types::{DataType, ScalarImpl};
 
-use crate::catalog::system_catalog::SystemCatalogColumnsDef;
+use crate::catalog::system_catalog::{SysCatalogReaderImpl, SystemCatalogColumnsDef};
+use crate::user::user_authentication::encrypted_raw_password;
 
 /// The view `pg_shadow` exists for backwards compatibility: it emulates a catalog that existed in
 /// PostgreSQL before version 8.1. It shows properties of all roles that are marked as rolcanlogin
@@ -39,3 +43,46 @@ pub static PG_SHADOW_COLUMNS: LazyLock<Vec<SystemCatalogColumnsDef<'_>>> = LazyL
         (DataType::List(Box::new(DataType::Varchar)), "useconfig"),
     ]
 });
+
+impl SysCatalogReaderImpl {
+    pub fn read_user_info_shadow(&self) -> Result<Vec<OwnedRow>> {
+        let reader = self.user_info_reader.read_guard();
+        // Since this catalog contains passwords, it must not be publicly readable.
+        match reader.get_user_by_name(&self.auth_context.user_name) {
+            None => {
+                return Err(ErrorCode::CatalogError(
+                    format!("user {} not found", self.auth_context.user_name).into(),
+                )
+                .into());
+            }
+            Some(user) => {
+                if !user.is_super {
+                    return Err(ErrorCode::PermissionDenied(
+                        "permission denied for table pg_shadow".to_string(),
+                    )
+                    .into());
+                }
+            }
+        }
+
+        let users = reader.get_all_users();
+        Ok(users
+            .iter()
+            .map(|user| {
+                OwnedRow::new(vec![
+                    Some(ScalarImpl::Utf8(user.name.clone().into())),
+                    Some(ScalarImpl::Int32(user.id as i32)),
+                    Some(ScalarImpl::Bool(user.can_create_db)),
+                    Some(ScalarImpl::Bool(user.is_super)),
+                    Some(ScalarImpl::Bool(false)),
+                    Some(ScalarImpl::Bool(false)),
+                    user.auth_info
+                        .as_ref()
+                        .map(|info| ScalarImpl::Utf8(encrypted_raw_password(info).into())),
+                    None,
+                    None,
+                ])
+            })
+            .collect_vec())
+    }
+}
