+++ src/frontend/src/catalog/system_catalog/pg_catalog/pg_tables.rs
@@ -12,9 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use risingwave_common::types::DataType;
+use itertools::Itertools;
+use risingwave_common::error::Result;
+use risingwave_common::row::OwnedRow;
+use risingwave_common::types::{DataType, ScalarImpl};
 
-use crate::catalog::system_catalog::SystemCatalogColumnsDef;
+use crate::catalog::system_catalog::{SysCatalogReaderImpl, SystemCatalogColumnsDef};
 
 /// The view `pg_tables` provides access to useful information about each table in the database.
 /// Ref: [`https://www.postgresql.org/docs/current/view-pg-tables.html`]
@@ -27,3 +30,45 @@ pub const PG_TABLES_COLUMNS: &[SystemCatalogColumnsDef<'_>] = &[
     (DataType::Varchar, "tablespace"), /* Since we don't have any concept of tablespace, we will
                                         * set this to null. */
 ];
+
+impl SysCatalogReaderImpl {
+    pub fn read_pg_tables_info(&self) -> Result<Vec<OwnedRow>> {
+        // TODO: avoid acquire two read locks here. The order is the same as in `read_views_info`.
+        let reader = self.catalog_reader.read_guard();
+        let user_info_reader = self.user_info_reader.read_guard();
+        let schemas = reader.iter_schemas(&self.auth_context.database)?;
+
+        Ok(schemas
+            .flat_map(|schema| {
+                schema
+                    .iter_table()
+                    .map(|table| {
+                        OwnedRow::new(vec![
+                            Some(ScalarImpl::Utf8(schema.name().into())),
+                            Some(ScalarImpl::Utf8(table.name().into())),
+                            Some(ScalarImpl::Utf8(
+                                user_info_reader
+                                    .get_user_name_by_id(table.owner)
+                                    .unwrap()
+                                    .into(),
+                            )),
+                            None,
+                        ])
+                    })
+                    .chain(schema.iter_system_tables().map(|table| {
+                        OwnedRow::new(vec![
+                            Some(ScalarImpl::Utf8(schema.name().into())),
+                            Some(ScalarImpl::Utf8(table.name().into())),
+                            Some(ScalarImpl::Utf8(
+                                user_info_reader
+                                    .get_user_name_by_id(table.owner)
+                                    .unwrap()
+                                    .into(),
+                            )),
+                            None,
+                        ])
+                    }))
+            })
+            .collect_vec())
+    }
+}
