+++ src/frontend/src/catalog/system_catalog/pg_catalog/pg_views.rs
@@ -12,9 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use risingwave_common::types::DataType;
+use itertools::Itertools;
+use risingwave_common::error::Result;
+use risingwave_common::row::OwnedRow;
+use risingwave_common::types::{DataType, ScalarImpl};
 
-use crate::catalog::system_catalog::SystemCatalogColumnsDef;
+use crate::catalog::system_catalog::{SysCatalogReaderImpl, SystemCatalogColumnsDef};
 
 /// The view `pg_views` provides access to useful information about each view in the database.
 /// Ref: [`https://www.postgresql.org/docs/current/view-pg-views.html`]
@@ -27,3 +30,33 @@ pub const PG_VIEWS_COLUMNS: &[SystemCatalogColumnsDef<'_>] = &[
     (DataType::Varchar, "viewowner"),
     (DataType::Varchar, "definition"),
 ];
+
+impl SysCatalogReaderImpl {
+    pub fn read_views_info(&self) -> Result<Vec<OwnedRow>> {
+        // TODO(zehua): solve the deadlock problem.
+        // Get two read locks. The order must be the same as
+        // `FrontendObserverNode::handle_initialization_notification`.
+        let catalog_reader = self.catalog_reader.read_guard();
+        let user_info_reader = self.user_info_reader.read_guard();
+        let schemas = catalog_reader.iter_schemas(&self.auth_context.database)?;
+
+        Ok(schemas
+            .flat_map(|schema| {
+                schema.iter_view().map(|view| {
+                    OwnedRow::new(vec![
+                        Some(ScalarImpl::Utf8(schema.name().into())),
+                        Some(ScalarImpl::Utf8(view.name().into())),
+                        Some(ScalarImpl::Utf8(
+                            user_info_reader
+                                .get_user_name_by_id(view.owner)
+                                .unwrap()
+                                .into(),
+                        )),
+                        // TODO(zehua): may be not same as postgresql's "definition" column.
+                        Some(ScalarImpl::Utf8(view.sql.clone().into())),
+                    ])
+                })
+            })
+            .collect_vec())
+    }
+}
