+++ src/frontend/src/catalog/system_catalog/rw_catalog/rw_actors.rs
@@ -12,9 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use risingwave_common::types::DataType;
+use itertools::Itertools;
+use risingwave_common::error::Result;
+use risingwave_common::row::OwnedRow;
+use risingwave_common::types::{DataType, ScalarImpl};
 
-use crate::catalog::system_catalog::SystemCatalogColumnsDef;
+use crate::catalog::system_catalog::{SysCatalogReaderImpl, SystemCatalogColumnsDef};
 
 pub const RW_ACTORS_TABLE_NAME: &str = "rw_actors";
 
@@ -24,3 +27,21 @@ pub const RW_ACTORS_COLUMNS: &[SystemCatalogColumnsDef<'_>] = &[
     (DataType::Int32, "parallel_unit_id"),
     (DataType::Varchar, "status"),
 ];
+
+impl SysCatalogReaderImpl {
+    pub async fn read_rw_actor_states_info(&self) -> Result<Vec<OwnedRow>> {
+        let states = self.meta_client.list_actor_states().await?;
+
+        Ok(states
+            .into_iter()
+            .map(|state| {
+                OwnedRow::new(vec![
+                    Some(ScalarImpl::Int32(state.actor_id as i32)),
+                    Some(ScalarImpl::Int32(state.fragment_id as i32)),
+                    Some(ScalarImpl::Int32(state.parallel_unit_id as i32)),
+                    Some(ScalarImpl::Utf8(state.state().as_str_name().into())),
+                ])
+            })
+            .collect_vec())
+    }
+}
