+++ src/frontend/src/catalog/system_catalog/rw_catalog/rw_databases.rs
@@ -12,9 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use risingwave_common::types::DataType;
+use itertools::Itertools;
+use risingwave_common::error::Result;
+use risingwave_common::row::OwnedRow;
+use risingwave_common::types::{DataType, ScalarImpl};
+use risingwave_pb::user::grant_privilege::Object;
 
-use crate::catalog::system_catalog::SystemCatalogColumnsDef;
+use crate::catalog::system_catalog::{
+    get_acl_items, SysCatalogReaderImpl, SystemCatalogColumnsDef,
+};
 
 pub const RW_DATABASES_TABLE_NAME: &str = "rw_databases";
 
@@ -24,3 +30,26 @@ pub const RW_DATABASES_COLUMNS: &[SystemCatalogColumnsDef<'_>] = &[
     (DataType::Int32, "owner"),
     (DataType::Varchar, "acl"),
 ];
+
+impl SysCatalogReaderImpl {
+    pub fn read_rw_database_info(&self) -> Result<Vec<OwnedRow>> {
+        let reader = self.catalog_reader.read_guard();
+        let user_reader = self.user_info_reader.read_guard();
+        let users = user_reader.get_all_users();
+        let username_map = user_reader.get_user_name_map();
+
+        Ok(reader
+            .iter_databases()
+            .map(|db| {
+                OwnedRow::new(vec![
+                    Some(ScalarImpl::Int32(db.id() as i32)),
+                    Some(ScalarImpl::Utf8(db.name().into())),
+                    Some(ScalarImpl::Int32(db.owner() as i32)),
+                    Some(ScalarImpl::Utf8(
+                        get_acl_items(&Object::DatabaseId(db.id()), &users, username_map).into(),
+                    )),
+                ])
+            })
+            .collect_vec())
+    }
+}
