+++ src/frontend/src/catalog/system_catalog/rw_catalog/rw_ddl_progress.rs
@@ -12,9 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use risingwave_common::types::DataType;
+use itertools::Itertools;
+use risingwave_common::error::Result;
+use risingwave_common::row::OwnedRow;
+use risingwave_common::types::{DataType, ScalarImpl};
 
-use crate::catalog::system_catalog::SystemCatalogColumnsDef;
+use crate::catalog::system_catalog::{SysCatalogReaderImpl, SystemCatalogColumnsDef};
 
 pub const RW_DDL_PROGRESS_TABLE_NAME: &str = "rw_ddl_progress";
 
@@ -23,3 +26,22 @@ pub const RW_DDL_PROGRESS_COLUMNS: &[SystemCatalogColumnsDef<'_>] = &[
     (DataType::Varchar, "ddl_statement"),
     (DataType::Varchar, "progress"),
 ];
+
+impl SysCatalogReaderImpl {
+    pub async fn read_ddl_progress(&self) -> Result<Vec<OwnedRow>> {
+        let ddl_progress = self
+            .meta_client
+            .list_ddl_progress()
+            .await?
+            .into_iter()
+            .map(|s| {
+                OwnedRow::new(vec![
+                    Some(ScalarImpl::Int64(s.id as i64)),
+                    Some(ScalarImpl::Utf8(s.statement.into())),
+                    Some(ScalarImpl::Utf8(s.progress.into())),
+                ])
+            })
+            .collect_vec();
+        Ok(ddl_progress)
+    }
+}
