+++ src/frontend/src/catalog/system_catalog/rw_catalog/rw_indexes.rs
@@ -12,9 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use risingwave_common::types::DataType;
+use itertools::Itertools;
+use risingwave_common::error::Result;
+use risingwave_common::row::OwnedRow;
+use risingwave_common::types::{DataType, ScalarImpl};
 
-use crate::catalog::system_catalog::SystemCatalogColumnsDef;
+use crate::catalog::system_catalog::{SysCatalogReaderImpl, SystemCatalogColumnsDef};
 
 pub const RW_INDEXES_TABLE_NAME: &str = "rw_indexes";
 
@@ -26,3 +29,25 @@ pub const RW_INDEXES_COLUMNS: &[SystemCatalogColumnsDef<'_>] = &[
     (DataType::Varchar, "definition"),
     (DataType::Varchar, "acl"),
 ];
+
+impl SysCatalogReaderImpl {
+    pub fn read_rw_indexes_info(&self) -> Result<Vec<OwnedRow>> {
+        let reader = self.catalog_reader.read_guard();
+        let schemas = reader.iter_schemas(&self.auth_context.database)?;
+
+        Ok(schemas
+            .flat_map(|schema| {
+                schema.iter_index().map(|index| {
+                    OwnedRow::new(vec![
+                        Some(ScalarImpl::Int32(index.id.index_id as i32)),
+                        Some(ScalarImpl::Utf8(index.name.clone().into())),
+                        Some(ScalarImpl::Int32(schema.id() as i32)),
+                        Some(ScalarImpl::Int32(index.index_table.owner as i32)),
+                        Some(ScalarImpl::Utf8(index.index_table.create_sql().into())),
+                        Some(ScalarImpl::Utf8("".into())),
+                    ])
+                })
+            })
+            .collect_vec())
+    }
+}
