+++ src/frontend/src/catalog/system_catalog/rw_catalog/rw_meta_snapshot.rs
@@ -12,9 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use risingwave_common::types::DataType;
+use itertools::Itertools;
+use risingwave_common::error::Result;
+use risingwave_common::row::OwnedRow;
+use risingwave_common::types::{DataType, ScalarImpl, Timestamp};
+use risingwave_common::util::epoch::Epoch;
 
-use crate::catalog::system_catalog::SystemCatalogColumnsDef;
+use crate::catalog::system_catalog::{SysCatalogReaderImpl, SystemCatalogColumnsDef};
 
 pub const RW_META_SNAPSHOT_TABLE_NAME: &str = "rw_meta_snapshot";
 
@@ -30,3 +34,37 @@ pub const RW_META_SNAPSHOT_COLUMNS: &[SystemCatalogColumnsDef<'_>] = &[
     // human-readable timestamp of max_committed_epoch
     (DataType::Timestamp, "max_committed_epoch_ts"),
 ];
+
+impl SysCatalogReaderImpl {
+    pub async fn read_meta_snapshot(&self) -> Result<Vec<OwnedRow>> {
+        let try_get_date_time = |epoch: u64| {
+            if epoch == 0 {
+                return None;
+            }
+            let time_millis = Epoch::from(epoch).as_unix_millis();
+            Timestamp::with_secs_nsecs(
+                (time_millis / 1000) as i64,
+                (time_millis % 1000 * 1_000_000) as u32,
+            )
+            .map(ScalarImpl::Timestamp)
+            .ok()
+        };
+        let meta_snapshots = self
+            .meta_client
+            .list_meta_snapshots()
+            .await?
+            .into_iter()
+            .map(|s| {
+                OwnedRow::new(vec![
+                    Some(ScalarImpl::Int64(s.id as i64)),
+                    Some(ScalarImpl::Int64(s.hummock_version_id as i64)),
+                    Some(ScalarImpl::Int64(s.safe_epoch as i64)),
+                    try_get_date_time(s.safe_epoch),
+                    Some(ScalarImpl::Int64(s.max_committed_epoch as i64)),
+                    try_get_date_time(s.max_committed_epoch),
+                ])
+            })
+            .collect_vec();
+        Ok(meta_snapshots)
+    }
+}
