+++ src/frontend/src/catalog/system_catalog/rw_catalog/rw_parallel_units.rs
@@ -12,11 +12,32 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use risingwave_common::types::DataType;
+use itertools::Itertools;
+use risingwave_common::error::Result;
+use risingwave_common::row::OwnedRow;
+use risingwave_common::types::{DataType, ScalarImpl};
 
-use crate::catalog::system_catalog::SystemCatalogColumnsDef;
+use crate::catalog::system_catalog::{SysCatalogReaderImpl, SystemCatalogColumnsDef};
 
 pub const RW_PARALLEL_UNITS_TABLE_NAME: &str = "rw_parallel_units";
 
 pub const RW_PARALLEL_UNITS_COLUMNS: &[SystemCatalogColumnsDef<'_>] =
     &[(DataType::Int32, "id"), (DataType::Int32, "worker_id")];
+
+impl SysCatalogReaderImpl {
+    pub fn read_rw_parallel_units_info(&self) -> Result<Vec<OwnedRow>> {
+        let workers = self.worker_node_manager.list_worker_nodes();
+
+        Ok(workers
+            .into_iter()
+            .flat_map(|worker| {
+                worker.parallel_units.into_iter().map(|unit| {
+                    OwnedRow::new(vec![
+                        Some(ScalarImpl::Int32(unit.id as i32)),
+                        Some(ScalarImpl::Int32(unit.worker_node_id as i32)),
+                    ])
+                })
+            })
+            .collect_vec())
+    }
+}
