+++ src/frontend/src/catalog/system_catalog/rw_catalog/rw_schemas.rs
@@ -12,9 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use risingwave_common::types::DataType;
+use itertools::Itertools;
+use risingwave_common::error::Result;
+use risingwave_common::row::OwnedRow;
+use risingwave_common::types::{DataType, ScalarImpl};
+use risingwave_pb::user::grant_privilege::Object;
 
-use crate::catalog::system_catalog::SystemCatalogColumnsDef;
+use crate::catalog::system_catalog::{
+    get_acl_items, SysCatalogReaderImpl, SystemCatalogColumnsDef,
+};
 
 pub const RW_SCHEMAS_TABLE_NAME: &str = "rw_schemas";
 
@@ -24,3 +30,26 @@ pub const RW_SCHEMAS_COLUMNS: &[SystemCatalogColumnsDef<'_>] = &[
     (DataType::Int32, "owner"),
     (DataType::Varchar, "acl"),
 ];
+
+impl SysCatalogReaderImpl {
+    pub fn read_rw_schema_info(&self) -> Result<Vec<OwnedRow>> {
+        let reader = self.catalog_reader.read_guard();
+        let schemas = reader.iter_schemas(&self.auth_context.database)?;
+        let user_reader = self.user_info_reader.read_guard();
+        let users = user_reader.get_all_users();
+        let username_map = user_reader.get_user_name_map();
+
+        Ok(schemas
+            .map(|schema| {
+                OwnedRow::new(vec![
+                    Some(ScalarImpl::Int32(schema.id() as i32)),
+                    Some(ScalarImpl::Utf8(schema.name().into())),
+                    Some(ScalarImpl::Int32(schema.owner() as i32)),
+                    Some(ScalarImpl::Utf8(
+                        get_acl_items(&Object::SchemaId(schema.id()), &users, username_map).into(),
+                    )),
+                ])
+            })
+            .collect_vec())
+    }
+}
