+++ src/frontend/src/catalog/system_catalog/rw_catalog/rw_sinks.rs
@@ -12,9 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use risingwave_common::types::DataType;
+use itertools::Itertools;
+use risingwave_common::error::Result;
+use risingwave_common::row::OwnedRow;
+use risingwave_common::types::{DataType, ScalarImpl};
+use risingwave_pb::user::grant_privilege::Object;
 
-use crate::catalog::system_catalog::SystemCatalogColumnsDef;
+use crate::catalog::system_catalog::{
+    get_acl_items, SysCatalogReaderImpl, SystemCatalogColumnsDef,
+};
+use crate::handler::create_source::UPSTREAM_SOURCE_KEY;
 
 pub const RW_SINKS_TABLE_NAME: &str = "rw_sinks";
 
@@ -29,3 +36,44 @@ pub const RW_SINKS_COLUMNS: &[SystemCatalogColumnsDef<'_>] = &[
     (DataType::Varchar, "definition"),
     (DataType::Varchar, "acl"),
 ];
+
+impl SysCatalogReaderImpl {
+    pub fn read_rw_sinks_info(&self) -> Result<Vec<OwnedRow>> {
+        let reader = self.catalog_reader.read_guard();
+        let schemas = reader.iter_schemas(&self.auth_context.database)?;
+        let user_reader = self.user_info_reader.read_guard();
+        let users = user_reader.get_all_users();
+        let username_map = user_reader.get_user_name_map();
+
+        Ok(schemas
+            .flat_map(|schema| {
+                schema.iter_sink().map(|sink| {
+                    OwnedRow::new(vec![
+                        Some(ScalarImpl::Int32(sink.id.sink_id as i32)),
+                        Some(ScalarImpl::Utf8(sink.name.clone().into())),
+                        Some(ScalarImpl::Int32(schema.id() as i32)),
+                        Some(ScalarImpl::Int32(sink.owner.user_id as i32)),
+                        Some(ScalarImpl::Utf8(
+                            sink.properties
+                                .get(UPSTREAM_SOURCE_KEY)
+                                .cloned()
+                                .unwrap_or("".to_string())
+                                .to_uppercase()
+                                .into(),
+                        )),
+                        Some(ScalarImpl::Utf8(
+                            sink.sink_type.to_proto().as_str_name().into(),
+                        )),
+                        sink.connection_id
+                            .map(|id| ScalarImpl::Int32(id.connection_id() as i32)),
+                        Some(ScalarImpl::Utf8(sink.create_sql().into())),
+                        Some(
+                            get_acl_items(&Object::SinkId(sink.id.sink_id), &users, username_map)
+                                .into(),
+                        ),
+                    ])
+                })
+            })
+            .collect_vec())
+    }
+}
