+++ src/frontend/src/catalog/system_catalog/rw_catalog/rw_sources.rs
@@ -14,9 +14,17 @@
 
 use std::sync::LazyLock;
 
-use risingwave_common::types::DataType;
+use itertools::Itertools;
+use risingwave_common::array::ListValue;
+use risingwave_common::error::Result;
+use risingwave_common::row::OwnedRow;
+use risingwave_common::types::{DataType, ScalarImpl};
+use risingwave_pb::user::grant_privilege::Object;
 
-use crate::catalog::system_catalog::SystemCatalogColumnsDef;
+use crate::catalog::system_catalog::{
+    get_acl_items, SysCatalogReaderImpl, SystemCatalogColumnsDef,
+};
+use crate::handler::create_source::UPSTREAM_SOURCE_KEY;
 
 pub const RW_SOURCES_TABLE_NAME: &str = "rw_sources";
 
@@ -36,3 +44,55 @@ pub static RW_SOURCES_COLUMNS: LazyLock<Vec<SystemCatalogColumnsDef<'_>>> = Lazy
         (DataType::Varchar, "acl"),
     ]
 });
+
+impl SysCatalogReaderImpl {
+    pub fn read_rw_sources_info(&self) -> Result<Vec<OwnedRow>> {
+        let reader = self.catalog_reader.read_guard();
+        let schemas = reader.iter_schemas(&self.auth_context.database)?;
+        let user_reader = self.user_info_reader.read_guard();
+        let users = user_reader.get_all_users();
+        let username_map = user_reader.get_user_name_map();
+
+        Ok(schemas
+            .flat_map(|schema| {
+                schema
+                    .iter_source()
+                    .filter(|s| s.associated_table_id.is_none())
+                    .map(|source| {
+                        OwnedRow::new(vec![
+                            Some(ScalarImpl::Int32(source.id as i32)),
+                            Some(ScalarImpl::Utf8(source.name.clone().into())),
+                            Some(ScalarImpl::Int32(schema.id() as i32)),
+                            Some(ScalarImpl::Int32(source.owner as i32)),
+                            Some(ScalarImpl::Utf8(
+                                source
+                                    .properties
+                                    .get(UPSTREAM_SOURCE_KEY)
+                                    .cloned()
+                                    .unwrap_or("".to_string())
+                                    .to_uppercase()
+                                    .into(),
+                            )),
+                            Some(ScalarImpl::List(ListValue::new(
+                                source
+                                    .columns
+                                    .iter()
+                                    .map(|c| Some(ScalarImpl::Utf8(c.name().into())))
+                                    .collect_vec(),
+                            ))),
+                            Some(ScalarImpl::Utf8(
+                                source.info.get_row_format().unwrap().as_str_name().into(),
+                            )),
+                            Some(ScalarImpl::Bool(source.append_only)),
+                            source.connection_id.map(|id| ScalarImpl::Int32(id as i32)),
+                            Some(ScalarImpl::Utf8(source.create_sql().into())),
+                            Some(
+                                get_acl_items(&Object::SourceId(source.id), &users, username_map)
+                                    .into(),
+                            ),
+                        ])
+                    })
+            })
+            .collect_vec())
+    }
+}
