+++ src/frontend/src/catalog/system_catalog/rw_catalog/rw_table_stats.rs
@@ -12,9 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use risingwave_common::types::DataType;
+use risingwave_common::error::Result;
+use risingwave_common::row::OwnedRow;
+use risingwave_common::types::{DataType, ScalarImpl};
 
-use crate::catalog::system_catalog::SystemCatalogColumnsDef;
+use crate::catalog::system_catalog::{SysCatalogReaderImpl, SystemCatalogColumnsDef};
 
 pub const RW_TABLE_STATS_TABLE_NAME: &str = "rw_table_stats";
 pub const RW_TABLE_STATS_TABLE_ID_INDEX: usize = 0;
@@ -27,3 +29,20 @@ pub const RW_TABLE_STATS_COLUMNS: &[SystemCatalogColumnsDef<'_>] = &[
     (DataType::Int64, "total_key_size"),
     (DataType::Int64, "total_value_size"),
 ];
+
+impl SysCatalogReaderImpl {
+    pub fn read_table_stats(&self) -> Result<Vec<OwnedRow>> {
+        let catalog = self.catalog_reader.read_guard();
+        let table_stats = catalog.table_stats();
+        let mut rows = vec![];
+        for (id, stats) in &table_stats.table_stats {
+            rows.push(OwnedRow::new(vec![
+                Some(ScalarImpl::Int32(*id as i32)),
+                Some(ScalarImpl::Int64(stats.total_key_count)),
+                Some(ScalarImpl::Int64(stats.total_key_size)),
+                Some(ScalarImpl::Int64(stats.total_value_size)),
+            ]));
+        }
+        Ok(rows)
+    }
+}
