+++ src/frontend/src/catalog/system_catalog/rw_catalog/rw_views.rs
@@ -12,9 +12,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use risingwave_common::types::DataType;
+use itertools::Itertools;
+use risingwave_common::error::Result;
+use risingwave_common::row::OwnedRow;
+use risingwave_common::types::{DataType, ScalarImpl};
+use risingwave_pb::user::grant_privilege::Object;
 
-use crate::catalog::system_catalog::SystemCatalogColumnsDef;
+use crate::catalog::system_catalog::{
+    get_acl_items, SysCatalogReaderImpl, SystemCatalogColumnsDef,
+};
 
 pub const RW_VIEWS_TABLE_NAME: &str = "rw_views";
 
@@ -26,3 +32,30 @@ pub const RW_VIEWS_COLUMNS: &[SystemCatalogColumnsDef<'_>] = &[
     (DataType::Varchar, "definition"),
     (DataType::Varchar, "acl"),
 ];
+
+impl SysCatalogReaderImpl {
+    pub fn read_rw_views_info(&self) -> Result<Vec<OwnedRow>> {
+        let reader = self.catalog_reader.read_guard();
+        let schemas = reader.iter_schemas(&self.auth_context.database)?;
+        let user_reader = self.user_info_reader.read_guard();
+        let users = user_reader.get_all_users();
+        let username_map = user_reader.get_user_name_map();
+
+        Ok(schemas
+            .flat_map(|schema| {
+                schema.iter_view().map(|view| {
+                    OwnedRow::new(vec![
+                        Some(ScalarImpl::Int32(view.id as i32)),
+                        Some(ScalarImpl::Utf8(view.name().into())),
+                        Some(ScalarImpl::Int32(schema.id() as i32)),
+                        Some(ScalarImpl::Int32(view.owner as i32)),
+                        Some(ScalarImpl::Utf8(view.create_sql().into())),
+                        Some(ScalarImpl::Utf8(
+                            get_acl_items(&Object::ViewId(view.id), &users, username_map).into(),
+                        )),
+                    ])
+                })
+            })
+            .collect_vec())
+    }
+}
