+++ src/frontend/src/catalog/system_catalog/rw_catalog/rw_worker_nodes.rs
@@ -12,9 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use risingwave_common::types::DataType;
+use itertools::Itertools;
+use risingwave_common::error::Result;
+use risingwave_common::row::OwnedRow;
+use risingwave_common::types::{DataType, ScalarImpl};
 
-use crate::catalog::system_catalog::SystemCatalogColumnsDef;
+use crate::catalog::system_catalog::{SysCatalogReaderImpl, SystemCatalogColumnsDef};
 
 /// `rw_worker_nodes` contains all information about the compute nodes in the cluster.
 /// TODO: Add other type of nodes if necessary in the future.
@@ -31,3 +34,32 @@ pub const RW_WORKER_NODES_COLUMNS: &[SystemCatalogColumnsDef<'_>] = &[
     (DataType::Boolean, "is_serving"),
     (DataType::Boolean, "is_unschedulable"),
 ];
+
+impl SysCatalogReaderImpl {
+    pub fn read_rw_worker_nodes_info(&self) -> Result<Vec<OwnedRow>> {
+        let workers = self.worker_node_manager.list_worker_nodes();
+
+        Ok(workers
+            .into_iter()
+            .map(|worker| {
+                let host = worker.host.as_ref().unwrap();
+                let property = worker.property.as_ref().unwrap();
+                OwnedRow::new(vec![
+                    Some(ScalarImpl::Int32(worker.id as i32)),
+                    Some(ScalarImpl::Utf8(host.host.clone().into())),
+                    Some(ScalarImpl::Utf8(host.port.to_string().into())),
+                    Some(ScalarImpl::Utf8(
+                        worker.get_type().unwrap().as_str_name().into(),
+                    )),
+                    Some(ScalarImpl::Utf8(
+                        worker.get_state().unwrap().as_str_name().into(),
+                    )),
+                    Some(ScalarImpl::Int32(worker.parallel_units.len() as i32)),
+                    Some(ScalarImpl::Bool(property.is_streaming)),
+                    Some(ScalarImpl::Bool(property.is_serving)),
+                    Some(ScalarImpl::Bool(property.is_unschedulable)),
+                ])
+            })
+            .collect_vec())
+    }
+}
